+++ src/openeo_aggregator/backend.py
@@ -22,7 +22,7 @@ from openeo_aggregator.errors import BackendLookupFailureException
 from openeo_aggregator.partitionedjobs import PartitionedJob
 from openeo_aggregator.partitionedjobs.splitting import FlimsySplitter, TileGridSplitter
 from openeo_aggregator.partitionedjobs.tracking import PartitionedJobConnection, PartitionedJobTracker
-from openeo_aggregator.utils import TtlCache, MultiDictGetter, subdict, dict_merge
+from openeo_aggregator.utils import TtlCache, MultiDictGetter, subdict, dict_merge, normalize_issuer_url
 from openeo_driver.ProcessGraphDeserializer import SimpleProcessing
 from openeo_driver.backend import OpenEoBackendImplementation, AbstractCollectionCatalog, LoadParameters, Processing, \
     OidcProvider, BatchJobs, BatchJobMetadata
@@ -432,6 +432,10 @@ class AggregatorProcessing(Processing):
                         aggregator_job_id=arguments["id"]
                     )
                     backend_candidates = [b for b in backend_candidates if b == job_backend_id]
+                elif process_id == "load_ml_model":
+                    model_backend_id = self._process_load_ml_model(arguments)[0]
+                    if model_backend_id:
+                        backend_candidates = [b for b in backend_candidates if b == model_backend_id]
         except Exception as e:
             _log.error(f"Failed to parse process graph: {e!r}", exc_info=True)
             raise ProcessGraphInvalidException()
@@ -497,12 +501,32 @@ class AggregatorProcessing(Processing):
                         assert job_backend_id == backend_id, f"{job_backend_id} != {backend_id}"
                         # Create new load_result node dict with updated job id
                         return dict_merge(node, arguments=dict_merge(arguments, id=job_id))
+                    if process_id == "load_ml_model":
+                        model_id = self._process_load_ml_model(arguments, expected_backend=backend_id)[1]
+                        if model_id:
+                            return dict_merge(node, arguments=dict_merge(arguments, id=model_id))
                 return {k: preprocess(v) for k, v in node.items()}
             elif isinstance(node, list):
                 return [preprocess(x) for x in node]
             return node
 
         return preprocess(process_graph)
+
+    def _process_load_ml_model(
+            self, arguments: dict, expected_backend: Optional[str] = None
+    ) -> Tuple[Union[str, None], str]:
+        """Handle load_ml_model: detect/strip backend_id from model_id if it is a job_id"""
+        model_id = arguments.get("id")
+        if model_id and not model_id.startswith("http"):
+            # TODO: load_ml_model's `id` could also be file path (see https://github.com/Open-EO/openeo-processes/issues/384)
+            job_id, job_backend_id = JobIdMapping.parse_aggregator_job_id(
+                backends=self.backends,
+                aggregator_job_id=model_id
+            )
+            if expected_backend and job_backend_id != expected_backend:
+                raise BackendLookupFailureException(f"{job_backend_id} != {expected_backend}")
+            return job_backend_id, job_id
+        return None, model_id
 
 
 class AggregatorBatchJobs(BatchJobs):
@@ -735,6 +759,7 @@ class AggregatorBackendImplementation(OpenEoBackendImplementation):
         )
         self._cache = TtlCache(default_ttl=CACHE_TTL_DEFAULT, name="General")
         self._backends.on_connections_change.add(self._cache.flush_all)
+        self._configured_oidc_providers: List[OidcProvider] = config.configured_oidc_providers
         self._auth_entitlement_check: Union[bool, dict] = config.auth_entitlement_check
 
         # Shorter HTTP cache TTL to adapt quicker to changed back-end configurations
@@ -743,11 +768,7 @@ class AggregatorBackendImplementation(OpenEoBackendImplementation):
         )
 
     def oidc_providers(self) -> List[OidcProvider]:
-        # TODO: openeo-python-driver (HttpAuthHandler) currently does support changes in
-        #       the set of oidc_providers ids (id mapping is statically established at startup time)
-        return self._cache.get_or_call(
-            key="oidc_providers", callback=self._backends.get_oidc_providers, log_on_miss=True,
-        )
+        return self._configured_oidc_providers
 
     def file_formats(self) -> dict:
         return self._cache.get_or_call(key="file_formats", callback=self._file_formats, log_on_miss=True)
@@ -777,10 +798,13 @@ class AggregatorBackendImplementation(OpenEoBackendImplementation):
     def user_access_validation(self, user: User, request: flask.Request) -> User:
         if self._auth_entitlement_check:
             int_data = user.internal_auth_data
-            issuer_whitelist = self._auth_entitlement_check.get("oidc_issuer_whitelist", [])
+            issuer_whitelist = [
+                normalize_issuer_url(u)
+                for u in self._auth_entitlement_check.get("oidc_issuer_whitelist", [])
+            ]
             if not (
                     int_data["authentication_method"] == "OIDC"
-                    and int_data["oidc_issuer"].rstrip("/").lower() in issuer_whitelist
+                    and normalize_issuer_url(int_data["oidc_issuer"]) in issuer_whitelist
             ):
                 user_message = "An EGI account is required for using openEO Platform."
                 _log.warning(f"user_access_validation failure: %r %r", user_message, {