+++ tests/test_backend.py
@@ -13,44 +13,14 @@ from openeo_driver.users.oidc import OidcProvider
 class TestAggregatorBackendImplementation:
 
     def test_oidc_providers(self, multi_backend_connection, config, backend1, backend2, requests_mock):
-        requests_mock.get(backend1 + "/credentials/oidc", json={"providers": [
-            {"id": "x", "issuer": "https://x.test", "title": "X"},
-            {"id": "y", "issuer": "https://y.test", "title": "YY"},
-        ]})
-        requests_mock.get(backend2 + "/credentials/oidc", json={"providers": [
-            {"id": "y", "issuer": "https://y.test", "title": "YY"},
-            {"id": "z", "issuer": "https://z.test", "title": "ZZZ"},
-        ]})
         implementation = AggregatorBackendImplementation(backends=multi_backend_connection, config=config)
         providers = implementation.oidc_providers()
         assert providers == [
-            OidcProvider(id="y-agg", issuer="https://y.test", title="Y (agg)")
+            OidcProvider(id='egi', issuer='https://egi.test', title='EGI'),
+            OidcProvider(id='x-agg', issuer='https://x.test', title='X (agg)'),
+            OidcProvider(id='y-agg', issuer='https://y.test', title='Y (agg)'),
+            OidcProvider(id='z-agg', issuer='https://z.test', title='Z (agg)'),
         ]
-
-    def test_oidc_providers_caching(self, multi_backend_connection, config, backend1, backend2, requests_mock):
-        m1 = requests_mock.get(backend1 + "/credentials/oidc", json={"providers": [
-            {"id": "x", "issuer": "https://x.test", "title": "X"},
-            {"id": "y", "issuer": "https://y.test", "title": "YY"},
-        ]})
-        m2 = requests_mock.get(backend2 + "/credentials/oidc", json={"providers": [
-            {"id": "y", "issuer": "https://y.test", "title": "YY"},
-            {"id": "z", "issuer": "https://z.test", "title": "ZZZ"},
-        ]})
-        implementation = AggregatorBackendImplementation(backends=multi_backend_connection, config=config)
-        assert (m1.call_count, m2.call_count) == (0, 0)
-        providers = implementation.oidc_providers()
-        assert providers == [OidcProvider(id="y-agg", issuer="https://y.test", title="Y (agg)")]
-        assert (m1.call_count, m2.call_count) == (1, 1)
-        providers = implementation.oidc_providers()
-        assert providers == [OidcProvider(id="y-agg", issuer="https://y.test", title="Y (agg)")]
-        assert (m1.call_count, m2.call_count) == (1, 1)
-
-        MultiBackendConnection._clock = itertools.count(time.time() + 1000).__next__
-        implementation._cache.flush_all()
-
-        providers = implementation.oidc_providers()
-        assert providers == [OidcProvider(id="y-agg", issuer="https://y.test", title="Y (agg)")]
-        assert (m1.call_count, m2.call_count) == (2, 2)
 
     def test_file_formats_simple(self, multi_backend_connection, config, backend1, backend2, requests_mock):
         just_geotiff = {